move ClassNodeAnalyzer to core
