--- packages/anonymous-class/config/config.php
@@ -1,15 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-use Symfony\Component\DependencyInjection\Loader\Configurator\ContainerConfigurator;
-
-return static function (ContainerConfigurator $containerConfigurator): void {
-    $services = $containerConfigurator->services();
-
-    $services->defaults()
-        ->autowire()
-        ->public();
-
-    $services->load('Rector\AnonymousClass\\', __DIR__ . '/../src');
-};
